broeye/requirements.txt
absl-py==0.8.1
altgraph==0.17
appdirs==1.4.3
asgiref==3.2.3
astor==0.8.0
auto-py-to-exe==2.6.6
beautifulsoup4==4.8.2
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.15
distlib==0.3.0
Django==3.0.3
dlib==19.17.0
Eel==0.11.0
face-recognition==1.2.3
face-recognition-models==0.3.0
filelock==3.0.12
future==0.18.2
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.3
h5py==2.10.0
idna==2.8
image==1.5.28
importlib-metadata==1.5.0
imutils==0.5.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
macholib==1.14
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.1
oauthlib==3.1.0
opencv-contrib-python==4.0.0.21
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.24.1
Pillow==6.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio==0.2.11
PyInstaller==3.6
pyobjc==6.1
pyobjc-core==6.1
pyobjc-framework-Accounts==6.1
pyobjc-framework-AddressBook==6.1
pyobjc-framework-AppleScriptKit==6.1
pyobjc-framework-AppleScriptObjC==6.1
pyobjc-framework-ApplicationServices==6.1
pyobjc-framework-Automator==6.1
pyobjc-framework-AVFoundation==6.1
pyobjc-framework-AVKit==6.1
pyobjc-framework-CalendarStore==6.1
pyobjc-framework-CFNetwork==6.1
pyobjc-framework-CloudKit==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-Collaboration==6.1
pyobjc-framework-ColorSync==6.1
pyobjc-framework-Contacts==6.1
pyobjc-framework-ContactsUI==6.1
pyobjc-framework-CoreAudio==6.1
pyobjc-framework-CoreAudioKit==6.1
pyobjc-framework-CoreBluetooth==6.1
pyobjc-framework-CoreData==6.1
pyobjc-framework-CoreLocation==6.1
pyobjc-framework-CoreMedia==6.1
pyobjc-framework-CoreMediaIO==6.1
pyobjc-framework-CoreML==6.1
pyobjc-framework-CoreServices==6.1
pyobjc-framework-CoreSpotlight==6.1
pyobjc-framework-CoreText==6.1
pyobjc-framework-CoreWLAN==6.1
pyobjc-framework-CryptoTokenKit==6.1
pyobjc-framework-DictionaryServices==6.1
pyobjc-framework-DiscRecording==6.1
pyobjc-framework-DiscRecordingUI==6.1
pyobjc-framework-DiskArbitration==6.1
pyobjc-framework-DVDPlayback==6.1
pyobjc-framework-EventKit==6.1
pyobjc-framework-ExceptionHandling==6.1
pyobjc-framework-ExternalAccessory==6.1
pyobjc-framework-FinderSync==6.1
pyobjc-framework-FSEvents==6.1
pyobjc-framework-GameCenter==6.1
pyobjc-framework-GameController==6.1
pyobjc-framework-GameKit==6.1
pyobjc-framework-GameplayKit==6.1
pyobjc-framework-ImageCaptureCore==6.1
pyobjc-framework-IMServicePlugIn==6.1
pyobjc-framework-InputMethodKit==6.1
pyobjc-framework-InstallerPlugins==6.1
pyobjc-framework-InstantMessage==6.1
pyobjc-framework-Intents==6.1
pyobjc-framework-IOSurface==6.1
pyobjc-framework-iTunesLibrary==6.1
pyobjc-framework-LatentSemanticMapping==6.1
pyobjc-framework-LaunchServices==6.1
pyobjc-framework-libdispatch==6.1
pyobjc-framework-LocalAuthentication==6.1
pyobjc-framework-MapKit==6.1
pyobjc-framework-MediaAccessibility==6.1
pyobjc-framework-MediaLibrary==6.1
pyobjc-framework-MediaPlayer==6.1
pyobjc-framework-MediaToolbox==6.1
pyobjc-framework-MetalKit==6.1
pyobjc-framework-ModelIO==6.1
pyobjc-framework-MultipeerConnectivity==6.1
pyobjc-framework-NetFS==6.1
pyobjc-framework-NetworkExtension==6.1
pyobjc-framework-NotificationCenter==6.1
pyobjc-framework-OpenDirectory==6.1
pyobjc-framework-OSAKit==6.1
pyobjc-framework-Photos==6.1
pyobjc-framework-PhotosUI==6.1
pyobjc-framework-PreferencePanes==6.1
pyobjc-framework-PubSub==6.1
pyobjc-framework-QTKit==6.1
pyobjc-framework-Quartz==6.1
pyobjc-framework-SafariServices==6.1
pyobjc-framework-SceneKit==6.1
pyobjc-framework-ScreenSaver==6.1
pyobjc-framework-ScriptingBridge==6.1
pyobjc-framework-SearchKit==6.1
pyobjc-framework-Security==6.1
pyobjc-framework-SecurityFoundation==6.1
pyobjc-framework-SecurityInterface==6.1
pyobjc-framework-ServiceManagement==6.1
pyobjc-framework-Social==6.1
pyobjc-framework-SpriteKit==6.1
pyobjc-framework-StoreKit==6.1
pyobjc-framework-SyncServices==6.1
pyobjc-framework-SystemConfiguration==6.1
pyobjc-framework-VideoToolbox==6.1
pyobjc-framework-Vision==6.1
pyobjc-framework-WebKit==6.1
pyparsing==2.4.2
pytesseract==0.3.2
python-dateutil==2.8.0
pyttsx3==2.81
pytz==2018.9
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
six==1.12.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tesseract==0.1.3
tk==0.1.0
urllib3==1.25.6
virtualenv==20.0.7
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.11.2
zipp==3.0.0
